+++ example/lib/widgets/examples/examples.dart
@@ -14,3 +14,4 @@ export 'read_only_example_widget.dart';
 export 'minimal_example_widget.dart';
 export 'tickmark_positioning_example_widget.dart';
 export 'segment_display_example_widget.dart';
+export 'segment_edit_example_widget.dart';